apps/service-category-embeddings/python/requirements.txt
gensim>=4.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
numpy
